composer/requirements.txt
psycopg2-binary==2.9.9
python-decouple==3.8
python-magic==0.4.27
python-pptx==0.6.23
python-slugify==8.0.4
redis==5.0.6
six==1.16.0
sqlalchemy==2.0.31
